Crud entidade accounting implementado

+++ src/main/java/exercicioprova/exercicioprova/repository/AccountingRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import exercicioprova.exercicioprova.model.Accounting;
 
-public interface AccountingRepository extends JpaRepository<Accounting, Integer> {
+public interface AccountingRepository extends JpaRepository<Accounting, Long> {
 
 }
